Move simple functions and MergeStrategy into skrample.common
Good for low level objects that don't need any depends
